+++ src/odig_odoc.mli
@@ -12,9 +12,6 @@
 
 open Odig_support
 
-val set_theme : Conf.t -> B0_odoc.Theme.t -> (unit, string) result
-(** [set_theme t] sets theme [odoc_theme] in configuration [conf]. *)
-
 val gen :
   Conf.t -> force:bool -> index_title:string option ->
   index_intro:B0_std.Fpath.t option -> pkg_deps:bool -> tag_index:bool ->
@@ -32,6 +29,8 @@ val gen :
     {- [tag_index] if [true] a tag index is generated on the package list
        page and package pages hyperlink into it from the package information
        section.}} *)
+
+val install_theme : Conf.t -> B0_odoc.Theme.t option -> (unit, string) result
 
 (*---------------------------------------------------------------------------
    Copyright (c) 2018 The odig programmers